Update AboutPage and ContactPage components
Update the AboutPage component to include the correct name of the university the developer is currently attending. Also, update the ContactPage component to change the id of the form element from "msgUs" to "msgUsForm" for consistency.

Update accordion.css and contactPage.css

Remove unnecessary CSS rules from accordion.css and contactPage.css to optimize the code.